Complete the agent development under SpringBoot2.x version.
