+++ hippo4j-agent/config/agent.config
@@ -111,11 +111,11 @@ collector.is_resolve_dns_periodically=${SW_AGENT_COLLECTOR_IS_RESOLVE_DNS_PERIOD
 # Logging level
 logging.level=${SW_LOGGING_LEVEL:INFO}
 # Logging file_name
-logging.file_name=${SW_LOGGING_FILE_NAME:skywalking-api.log}
+logging.file_name=${SW_LOGGING_FILE_NAME:hippo4j-api.log}
 # Log output. Default is FILE. Use CONSOLE means output to stdout.
 logging.output=${SW_LOGGING_OUTPUT:FILE}
-# Log files directory. Default is blank string, meaning use "{theSkywalkingAgentJarDir}/logs " to output logs.
-# {theSkywalkingAgentJarDir} is the directory where the skywalking agent jar file is located
+# Log files directory. Default is blank string, meaning use "{theHippo4jAgentJarDir}/logs " to output logs.
+# {theHippo4jAgentJarDir} is the directory where the hippo4j agent jar file is located
 logging.dir=${SW_LOGGING_DIR:}
 # Logger resolver: PATTERN or JSON. The default is PATTERN, which uses logging.pattern to print traditional text logs.
 # JSON resolver prints logs in JSON format.
