+++ hippo4j-agent/hippo4j-agent-plugin/spring-plugins/spring-boot-1.x-plugin/src/main/java/cn/hippo4j/agent/plugin/spring/boot/v1/DynamicThreadPoolChangeHandlerSpring1x.java
@@ -1,0 +1,69 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cn.hippo4j.agent.plugin.spring.boot.v1;
+
+import cn.hippo4j.agent.plugin.spring.common.support.AbstractDynamicThreadPoolChangeHandlerSpring;
+import cn.hippo4j.common.toolkit.MapUtil;
+import cn.hippo4j.config.springboot.starter.config.BootstrapConfigProperties;
+import org.springframework.beans.PropertyValues;
+import org.springframework.beans.support.ResourceEditorRegistrar;
+import org.springframework.boot.bind.CustomPropertyNamePatternsMatcher;
+import org.springframework.boot.bind.RelaxedDataBinder;
+import org.springframework.boot.bind.RelaxedNames;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.core.env.MapPropertySource;
+import org.springframework.core.env.MutablePropertySources;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+
+import static cn.hippo4j.config.springboot1x.starter.refresher.SpringBoot1xBootstrapConfigPropertiesBinderAdapt.getNames;
+
+public class DynamicThreadPoolChangeHandlerSpring1x extends AbstractDynamicThreadPoolChangeHandlerSpring {
+
+    public DynamicThreadPoolChangeHandlerSpring1x(ConfigurableApplicationContext context) {
+        super(context);
+    }
+
+    protected BootstrapConfigProperties bindProperties(Map<Object, Object> configInfo, ApplicationContext applicationContext) {
+        BootstrapConfigProperties bindableCoreProperties = new BootstrapConfigProperties();
+        if (MapUtil.isEmpty(configInfo)) {
+            return bindableCoreProperties;
+        }
+        RelaxedNames relaxedNames = new RelaxedNames(BootstrapConfigProperties.PREFIX);
+        Set<String> names = getNames(bindableCoreProperties, relaxedNames);
+        Map<String, Object> stringConfigInfo = new HashMap<>(configInfo.size());
+        configInfo.forEach((key, value) -> stringConfigInfo.put(key.toString(), value));
+        MapPropertySource test = new MapPropertySource("Hippo4j", stringConfigInfo);
+        MutablePropertySources propertySources = new MutablePropertySources();
+        propertySources.addFirst(test);
+        PropertyValues propertyValues = CustomPropertyNamePatternsMatcher.getPropertySourcesPropertyValues(names, propertySources);
+        RelaxedDataBinder dataBinder = new RelaxedDataBinder(bindableCoreProperties, BootstrapConfigProperties.PREFIX);
+        dataBinder.setAutoGrowCollectionLimit(Integer.MAX_VALUE);
+        dataBinder.setIgnoreNestedProperties(false);
+        dataBinder.setIgnoreInvalidFields(false);
+        dataBinder.setIgnoreUnknownFields(true);
+        ResourceEditorRegistrar resourceEditorRegistrar = new ResourceEditorRegistrar(applicationContext, applicationContext.getEnvironment());
+        resourceEditorRegistrar.registerCustomEditors(dataBinder);
+        dataBinder.bind(propertyValues);
+        return bindableCoreProperties;
+    }
+
+}
