+++ hippo4j-agent/hippo4j-agent-plugin/spring-plugins/spring-boot-1.x-plugin/src/main/java/cn/hippo4j/agent/plugin/spring/boot/v1/interceptor/EventPublishingFinishedInterceptor.java
@@ -15,21 +15,27 @@
  * limitations under the License.
  */
 
-package cn.hippo4j.agent.plugin.spring.boot.v2;
+package cn.hippo4j.agent.plugin.spring.boot.v1.interceptor;
 
 import cn.hippo4j.agent.core.logging.api.ILog;
 import cn.hippo4j.agent.core.logging.api.LogManager;
 import cn.hippo4j.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import cn.hippo4j.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import cn.hippo4j.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import cn.hippo4j.agent.plugin.spring.common.SpringPropertiesLoader;
+import cn.hippo4j.agent.plugin.spring.boot.v1.DynamicThreadPoolChangeHandlerSpring1x;
+import cn.hippo4j.agent.plugin.spring.common.support.SpringPropertiesLoader;
+import cn.hippo4j.agent.plugin.spring.common.support.IDynamicThreadPoolChangeHandlerSpring;
+import cn.hippo4j.agent.plugin.spring.common.support.SpringThreadPoolRegisterSupport;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.ConfigurableApplicationContext;
 
 import java.lang.reflect.Method;
 
-public class EventPublishingStartedInterceptor implements InstanceMethodsAroundInterceptor {
+public class EventPublishingFinishedInterceptor implements InstanceMethodsAroundInterceptor {
 
-    private static final ILog LOGGER = LogManager.getLogger(EventPublishingStartedInterceptor.class);
+    private static final ILog FILE_LOGGER = LogManager.getLogger(EventPublishingFinishedInterceptor.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(EventPublishingFinishedInterceptor.class);
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes, MethodInterceptResult result) throws Throwable {
@@ -39,7 +45,15 @@ public class EventPublishingStartedInterceptor implements InstanceMethodsAroundI
     @Override
     public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes, Object ret) throws Throwable {
         ConfigurableApplicationContext context = (ConfigurableApplicationContext) allArguments[0];
+        if (context.getParent() != null) {
+            // After the child container is started, the thread pool registration will be carried out
+            SpringThreadPoolRegisterSupport.registerThreadPoolInstances();
+            return ret;
+        }
         SpringPropertiesLoader.loadSpringProperties(context.getEnvironment());
+        IDynamicThreadPoolChangeHandlerSpring handlerSpring1x = new DynamicThreadPoolChangeHandlerSpring1x(context);
+        handlerSpring1x.registerApolloConfigHandler();
+
         return ret;
     }
