+++ hippo4j-agent/hippo4j-agent-plugin/spring-plugins/spring-boot-2.x-plugin/src/main/java/cn/hippo4j/agent/plugin/spring/boot/v2/define/EventPublishingRunListenerInstrumentation.java
@@ -31,7 +31,8 @@ public class EventPublishingRunListenerInstrumentation extends ClassInstanceMeth
 
     private static final String ENHANCE_CLASS = "org.springframework.boot.context.event.EventPublishingRunListener";
 
-    private static final String EVENT_PUBLISHING_FINISHED_INTERCEPTOR = "cn.hippo4j.agent.plugin.spring.boot.v2.EventPublishingStartedInterceptor";
+    private static final String EVENT_PUBLISHING_FINISHED_INTERCEPTOR = "cn.hippo4j.agent.plugin.spring.boot.v2.interceptor.EventPublishingStartedInterceptor";
+    private static final String EVENT_PUBLISHING_ENVIRONMENT_PREPARED_INTERCEPTOR = "cn.hippo4j.agent.plugin.spring.common.interceptor.EventPublishingRunListenerEnvironmentPreparedInterceptor";
 
     @Override
     protected ClassMatch enhanceClass() {
@@ -56,6 +57,23 @@ public class EventPublishingRunListenerInstrumentation extends ClassInstanceMeth
                     @Override
                     public String getMethodsInterceptor() {
                         return EVENT_PUBLISHING_FINISHED_INTERCEPTOR;
+                    }
+
+                    @Override
+                    public boolean isOverrideArgs() {
+                        return false;
+                    }
+                },
+                new InstanceMethodsInterceptPoint() {
+
+                    @Override
+                    public ElementMatcher<MethodDescription> getMethodsMatcher() {
+                        return named("environmentPrepared");
+                    }
+
+                    @Override
+                    public String getMethodsInterceptor() {
+                        return EVENT_PUBLISHING_ENVIRONMENT_PREPARED_INTERCEPTOR;
                     }
 
                     @Override
